Remove test_model_func.py module
